Разделён service layer партнёрских программ и расширены тесты
