+++ docs/modules/partner-programs.md
@@ -36,7 +36,7 @@
   фильтров и экспортов.
 - `partner_programs/serializers/` - request/response serializers и validation
   дополнительных полей.
-- `partner_programs/services.py` - сервисы регистрации в программе, подачи
+- `partner_programs/services/` - сервисы регистрации в программе, подачи
   проектов, фильтрации проектов, публикации проектов и подготовки
   Excel-выгрузок.
 - `partner_programs/selectors.py` - выборки участников для аналитики и
@@ -205,4 +205,8 @@
 - список проектов программы для менеджера;
 - Excel-выгрузку проектов программы, включая режим `only_submitted`;
 - запрет выгрузки проектов пользователем без прав менеджера;
+- Excel-выгрузку оценок проектов программы;
+- подготовку данных выгрузки оценок, когда критерии есть, но оценок еще нет;
+- базовые permissions для менеджера программы, staff-пользователя,
+  постороннего пользователя, anonymous и лидера проекта;
 - публикацию проектов после завершения программы.
