+++ partner_programs/services/exports.py
@@ -1,17 +1,13 @@
-import logging
 import io
+import logging
 from collections import OrderedDict
 from dataclasses import dataclass
 
-from django.contrib.auth import get_user_model
-from django.db import IntegrityError, transaction
-from django.db.models import Exists, OuterRef, Prefetch
+from django.db.models import Prefetch
 from django.utils import timezone
 from openpyxl import Workbook
-from rest_framework.exceptions import PermissionDenied, ValidationError
 
 from core.utils import XlsxFileToExport, sanitize_excel_value
-from partner_programs.helpers import date_to_iso
 from partner_programs.models import (
     PartnerProgram,
     PartnerProgramField,
@@ -20,368 +16,15 @@ from partner_programs.models import (
     PartnerProgramUserProfile,
 )
 from project_rates.models import Criteria, ProjectScore
-from projects.models import Collaborator, Project
-from vacancy.mapping import MessageTypeEnum, UserProgramRegisterParams
-from vacancy.tasks import send_email
+from projects.models import Collaborator
 
 logger = logging.getLogger()
-User = get_user_model()
-
-EXTERNAL_REGISTRATION_USER_FIELDS = (
-    "first_name",
-    "last_name",
-    "patronymic",
-    "city",
-)
-
-
-class ProgramRegistrationError(Exception):
-    def __init__(self, detail: str):
-        self.detail = detail
-        super().__init__(detail)
-
-
-class ProgramProjectAlreadyApplied(Exception):
-    def __init__(self, program_link: PartnerProgramProject):
-        self.program_link = program_link
-        super().__init__("Проект уже подан в эту программу.")
-
-
-class ProgramProjectFilterError(Exception):
-    def __init__(self, detail: dict):
-        self.detail = detail
-        super().__init__(str(detail))
-
-
-@dataclass(frozen=True)
-class ProgramProjectApplicationResult:
-    project: Project
-    program_link: PartnerProgramProject
 
 
 @dataclass(frozen=True)
 class ProgramExportFile:
     binary_data: bytes
     base_name: str
-
-
-def _send_program_registration_email(user, program: PartnerProgram) -> None:
-    send_email.delay(
-        UserProgramRegisterParams(
-            message_type=MessageTypeEnum.REGISTERED_PROGRAM_USER.value,
-            user_id=user.id,
-            program_name=program.name,
-            program_id=program.id,
-            schema_id=2,
-        )
-    )
-
-
-def register_user_to_program(
-    *,
-    program: PartnerProgram,
-    user: User,
-    data,
-) -> PartnerProgramUserProfile:
-    if program.datetime_registration_ends < timezone.now():
-        raise ProgramRegistrationError("Registration period has ended.")
-
-    try:
-        user_profile = PartnerProgramUserProfile.objects.create(
-            partner_program_data=data,
-            user=user,
-            partner_program=program,
-        )
-    except IntegrityError:
-        raise ProgramRegistrationError("User already registered to this program.")
-
-    _send_program_registration_email(user, program)
-    return user_profile
-
-
-def create_user_and_register_to_program(
-    *,
-    program: PartnerProgram,
-    data,
-) -> PartnerProgramUserProfile:
-    email = data.get("email") if data.get("email") else data.get("email_")
-    if not email:
-        raise ProgramRegistrationError("You need to pass an email address.")
-
-    password = data.get("password")
-    if not password:
-        raise ProgramRegistrationError("You need to pass a password.")
-
-    user, created = User.objects.get_or_create(
-        email=email,
-        defaults={
-            "birthday": date_to_iso(data.get("birthday", "01-01-1900")),
-            "is_active": True,  # bypass email verification for external forms
-            "onboarding_stage": None,  # bypass onboarding for external forms
-            "verification_date": timezone.now(),  # bypass manual verification
-            **{
-                field_name: data.get(field_name, "")
-                for field_name in EXTERNAL_REGISTRATION_USER_FIELDS
-            },
-        },
-    )
-    if created:
-        user.set_password(password)
-        user.save()
-
-    user_profile_program_data = {
-        k: v
-        for k, v in data.items()
-        if k not in EXTERNAL_REGISTRATION_USER_FIELDS and k != "password"
-    }
-    try:
-        user_profile = PartnerProgramUserProfile.objects.create(
-            partner_program_data=user_profile_program_data,
-            user=user,
-            partner_program=program,
-        )
-    except IntegrityError:
-        raise ProgramRegistrationError(
-            "User has already registered in this program."
-        )
-
-    _send_program_registration_email(user, program)
-    return user_profile
-
-
-def require_can_apply_project_to_program(
-    *,
-    program: PartnerProgram,
-    user: User,
-) -> None:
-    if not program.is_project_submission_open():
-        raise ValidationError("Срок подачи проектов в программу завершён.")
-
-    if program.is_manager(user):
-        return
-
-    if not PartnerProgramUserProfile.objects.filter(
-        user=user,
-        partner_program=program,
-    ).exists():
-        raise PermissionDenied("Подача проекта доступна только участникам программы.")
-
-
-def _validate_unique_program_fields(values_data: list[dict]) -> None:
-    seen_field_ids: set[int] = set()
-    duplicate_ids: set[int] = set()
-    for item in values_data:
-        field_id = item["field"].id
-        if field_id in seen_field_ids:
-            duplicate_ids.add(field_id)
-        seen_field_ids.add(field_id)
-    if duplicate_ids:
-        raise ValidationError(
-            {"program_field_values": f"Есть повторяющиеся field_id: {sorted(duplicate_ids)}"}
-        )
-
-
-def _validate_required_program_fields(
-    *,
-    program: PartnerProgram,
-    values_data: list[dict],
-) -> None:
-    required_fields = list(
-        program.fields.filter(is_required=True).values("id", "label")
-    )
-    provided_field_ids = {item["field"].id for item in values_data}
-    missing_required = [
-        field["label"]
-        for field in required_fields
-        if field["id"] not in provided_field_ids
-    ]
-    if missing_required:
-        raise ValidationError(
-            {"program_field_values": f"Не заполнены обязательные поля: {missing_required}"}
-        )
-
-
-def _validate_program_field_ownership(
-    *,
-    program: PartnerProgram,
-    values_data: list[dict],
-) -> None:
-    for item in values_data:
-        field = item["field"]
-        if field.partner_program_id != program.id:
-            raise ValidationError(
-                {
-                    "program_field_values": f"Поле id={field.id} не относится к этой программе."
-                }
-            )
-
-
-def apply_project_to_program(
-    *,
-    program: PartnerProgram,
-    user: User,
-    data,
-    serializer_class,
-) -> ProgramProjectApplicationResult:
-    require_can_apply_project_to_program(program=program, user=user)
-
-    existing_link = (
-        PartnerProgramProject.objects.select_related("project")
-        .filter(partner_program=program, project__leader=user)
-        .first()
-    )
-    if existing_link:
-        raise ProgramProjectAlreadyApplied(existing_link)
-
-    serializer = serializer_class(data=data)
-    serializer.is_valid(raise_exception=True)
-    validated_data = serializer.validated_data
-
-    project_data = validated_data["project"]
-    values_data = validated_data.get("program_field_values") or []
-
-    _validate_unique_program_fields(values_data)
-    _validate_required_program_fields(program=program, values_data=values_data)
-    _validate_program_field_ownership(program=program, values_data=values_data)
-
-    with transaction.atomic():
-        project = Project.objects.create(
-            leader=user,
-            draft=True,
-            is_public=False,
-            **project_data,
-        )
-        program_link = PartnerProgramProject.objects.create(
-            partner_program=program,
-            project=project,
-        )
-
-        profile = PartnerProgramUserProfile.objects.filter(
-            user=user,
-            partner_program=program,
-        ).first()
-        if profile:
-            profile.project = project
-            profile.save(update_fields=["project"])
-
-        value_objs = [
-            PartnerProgramFieldValue(
-                program_project=program_link,
-                field=item["field"],
-                value_text=item.get("value_text") or "",
-            )
-            for item in values_data
-        ]
-        if value_objs:
-            PartnerProgramFieldValue.objects.bulk_create(value_objs)
-
-    return ProgramProjectApplicationResult(project=project, program_link=program_link)
-
-
-def get_filterable_program_fields(program: PartnerProgram):
-    return PartnerProgramField.objects.filter(
-        partner_program=program,
-        show_filter=True,
-    )
-
-
-def validate_program_project_filters(
-    *,
-    program: PartnerProgram,
-    filters: dict[str, list[str]],
-) -> None:
-    field_names = list(filters.keys())
-    field_qs = PartnerProgramField.objects.filter(
-        partner_program=program,
-        name__in=field_names,
-    )
-    field_by_name = {field.name: field for field in field_qs}
-
-    missing = [name for name in field_names if name not in field_by_name]
-    if missing:
-        raise ProgramProjectFilterError(
-            {"detail": f"Поля не найденные в программе: {missing}"}
-        )
-
-    for field_name, values in filters.items():
-        field_obj = field_by_name[field_name]
-        if not field_obj.show_filter:
-            raise ProgramProjectFilterError(
-                {
-                    "detail": (
-                        f"Поле '{field_name}' недоступно для фильтрации "
-                        "(show_filter=False)."
-                    )
-                }
-            )
-
-        options = field_obj.get_options_list()
-        if not options:
-            raise ProgramProjectFilterError(
-                {"detail": f"Поле '{field_name}' не имеет вариантов (options)."}
-            )
-
-        invalid_values = [value for value in values if value not in options]
-        if invalid_values:
-            raise ProgramProjectFilterError(
-                {
-                    "detail": f"Неверные значения для поля '{field_name}'.",
-                    "invalid": invalid_values,
-                }
-            )
-
-
-def get_filtered_program_project_links(
-    *,
-    program: PartnerProgram,
-    filters: dict[str, list[str]],
-):
-    validate_program_project_filters(program=program, filters=filters)
-
-    qs = PartnerProgramProject.objects.filter(partner_program=program)
-    if not filters:
-        return qs.select_related("project").distinct()
-
-    for field_name, values in filters.items():
-        field = PartnerProgramField.objects.get(
-            partner_program=program,
-            name=field_name.strip(),
-        )
-        field_value_exists = PartnerProgramFieldValue.objects.filter(
-            program_project=OuterRef("pk"),
-            field=field,
-            value_text__in=values,
-        )
-        qs = qs.filter(Exists(field_value_exists))
-
-    return qs.select_related("project").distinct()
-
-
-def publish_finished_program_projects(now=None) -> int:
-    if now is None:
-        now = timezone.now()
-
-    program_ids = PartnerProgram.objects.filter(
-        publish_projects_after_finish=True,
-        datetime_finished__lte=now,
-    ).values_list("id", flat=True)
-    if not program_ids.exists():
-        return 0
-
-    link_project_ids = PartnerProgramProject.objects.filter(
-        partner_program_id__in=program_ids
-    ).values_list("project_id", flat=True)
-    profile_project_ids = PartnerProgramUserProfile.objects.filter(
-        partner_program_id__in=program_ids,
-        project_id__isnull=False,
-    ).values_list("project_id", flat=True)
-    project_ids = link_project_ids.union(profile_project_ids)
-
-    return Project.objects.filter(
-        id__in=project_ids,
-        is_public=False,
-        draft=False,
-    ).update(is_public=True)
 
 
 class ProjectScoreDataPreparer:
